+++ scripts/download-sdks.ps1
@@ -34,7 +34,7 @@ if (-not (Test-Path $outDir))
     New-Item $outDir -ItemType Directory > $null
 }
 
-$sdks = @("Android", "IOS", "Linux", "LinuxArm64", "Mac", "Win64")
+$sdks = @("Android", "IOS", "Linux", "LinuxArm64", "Mac", "Win64", "WinArm64")
 foreach ($sdk in $sdks)
 {
     $sdkDir = "$outDir/$sdk"
@@ -50,5 +50,6 @@ foreach ($sdk in $sdks)
 
 Write-Host "Downloading Crash Reporter binaries ..."
 gh run download $runId -n "CrashReporter-Win64" -D "$outDir/Win64/bin"
+gh run download $runId -n "CrashReporter-WinArm64" -D "$outDir/WinArm64/bin"
 gh run download $runId -n "CrashReporter-Linux" -D "$outDir/Linux/bin"
 gh run download $runId -n "CrashReporter-LinuxArm64" -D "$outDir/LinuxArm64/bin"